1. Fixed some javadoc
2. Removed unused code
3. Removed use of streams
4. Removed Builder from AccessEvalutor for consistency
5. Added note in javadoc for API objects about UTF-8 expectations