@dlmarion 2024-01-25
```
main:

Benchmark                                           Mode  Cnt   Score   Error   Units
AccessExpressionBenchmark.measureBytesValidation   thrpt   12  32.489 ± 0.166  ops/us
AccessExpressionBenchmark.measureEvaluation        thrpt   12  12.720 ± 0.961  ops/us
AccessExpressionBenchmark.measureStringValidation  thrpt   12  25.196 ± 0.037  ops/us

this branch:

Benchmark                                           Mode  Cnt   Score   Error   Units
AccessExpressionBenchmark.measureBytesValidation   thrpt   12  32.560 ± 0.127  ops/us
AccessExpressionBenchmark.measureEvaluation        thrpt   12  13.292 ± 0.423  ops/us
AccessExpressionBenchmark.measureStringValidation  thrpt   12  24.660 ± 1.305  ops/us
```